requirements.txt
fastapi==0.75.1
pyserial-asyncio==0.6
uvicorn==0.17.6
websockets==10.2
watchgod==0.8.2
starlette==0.17.1

anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.0
pyserial==3.5
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
typing-extensions==4.1.1
